Removed fifth year transition paragraph
Removed the following paragraph from beginning of About section:
Minyan Dafna is entering its fifth year and is currently in a space of transition. Over the summer, we will continue to meet for Friday night davenings (the calendar will be updated soon; check our email list and Facebook page for updates). We would love to hear your input and suggestions for how you’d like to be involved and where you’d like to see Minyan Dafna go in the future.

B’Shalom,
The Minyan Dafna team